server/requirements.txt
Flask>=3.0
python-dotenv>=1.0
google-auth-oauthlib>=1.2
google-api-python-client>=2.141
beautifulsoup4>=4.12
lxml>=5.3
SQLAlchemy>=2.0
python-dateutil>=2.9
requests>=2.32
openai>=1.0
gunicorn>=21.2
